removed event parser
